+++ src/components/ui/AddLocation.js
@@ -3,6 +3,7 @@ import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { toast } from 'react-toastify'
+import { css } from 'styled-components'
 import styled from 'styled-components/macro'
 
 import { VISIBLE_CLUSTER_ZOOM_LIMIT } from '../../constants/map'
@@ -10,11 +11,7 @@ import { useAppHistory } from '../../utils/useAppHistory'
 import Button from './Button'
 import IconButton from './IconButton'
 
-const StyledAddLocationButton = styled(IconButton)`
-  position: absolute;
-  inset-block-end: calc(10px + var(--safe-area-inset-bottom, 0px));
-  inset-inline-end: 10px;
-  z-index: 1;
+const greyableButtonStyle = css`
   opacity: ${({ greyedOut, disabled }) =>
     greyedOut || disabled ? '0.5' : '1'};
   cursor: ${({ greyedOut, disabled }) => {
@@ -25,20 +22,21 @@ const StyledAddLocationButton = styled(IconButton)`
   }};
 `
 
+const StyledAddLocationButton = styled(IconButton)`
+  position: absolute;
+  inset-block-end: calc(10px + var(--safe-area-inset-bottom, 0px));
+  inset-inline-end: 10px;
+  z-index: 1;
+  ${greyableButtonStyle}
+`
+
 const AddLocationDesktopButton = styled(Button)`
   margin-inline: 0.75em;
   margin-block-start: 0;
   margin-block-end: 1em;
   padding-block: 1em;
   padding-inline: 0;
-  opacity: ${({ greyedOut, disabled }) =>
-    greyedOut || disabled ? '0.5' : '1'};
-  cursor: ${({ greyedOut, disabled }) => {
-    if (disabled) {
-      return 'not-allowed'
-    }
-    return greyedOut ? 'help' : 'pointer'
-  }};
+  ${greyableButtonStyle}
 `
 
 const useAddLocation = (addLocationPath) => {
@@ -49,7 +47,7 @@ const useAddLocation = (addLocationPath) => {
   const noMap = !googleMap
 
   const isZoomSufficient =
-    !googleMap || googleMap.getZoom() > VISIBLE_CLUSTER_ZOOM_LIMIT
+    !!googleMap && googleMap.getZoom() > VISIBLE_CLUSTER_ZOOM_LIMIT
 
   const handleAddLocation = () => {
     if (noMap) {
